Make time work in beats and not milliseconds
This changes pretty much everything that uses the time,
because it's now a float and not an int.
Also, the timelines and keyframes snap to whole numbers for now,
maybe that'll change later.

+++ src/CameraKeyframeController.cpp
@@ -18,7 +18,7 @@ void CameraKeyframeController::moveCameraToKeyframe()
     cameraController.markAsMoved();
 }
 
-void CameraKeyframeController::startFrame(long currentTime)
+void CameraKeyframeController::startFrame(float currentTime)
 {
     bool didTimeChange = this->currentTime != currentTime;
